refactor: give the two Kafka packages a shared home in _internal/kafka
`faststream/response/response.py` held ~95 lines whose only consumers were
`faststream/kafka/response.py` and `faststream/confluent/response.py`:
`extract_per_message_keys_and_bodies`, `key_for_index`, `realign_keys`, and the
`_extract_body_and_key` singledispatch. Per-message batch keys are a Kafka
concept, so a cross-broker module was the wrong place for them.

The two Kafka packages are deliberately independent -- neither imports the
other -- so neither could host the code either. Added
`faststream/_internal/kafka/` as the shared home, per the documented rule to
implement shared behavior in `_internal/` and expose it through broker packages.

`Response.get_publish_key()` stays on the base class: it returns `None` by
default and reads as a genuine extension point. The machinery moved, the hook
did not.

Pure move plus import rewiring -- no behavior change.
